python_integration/requirements.txt
<<<<<<< HEAD
logging==0.4.9.6
python-dateutil==2.5.3
requests==2.10.0
schedule==0.3.2

#Use the following command to install mysql-connector library
#easy_install mysql-connector==2.1.3
=======
requests==2.10.0; python_version == '3.4.1'
mysql-connector==2.1.3; python_version == '3.4.1'
schedule==0.3.2; python_version == '3.4.1'
>>>>>>> 302b4c28334a74bc731f6dd214df966dcb7b01d7
